requirements.txt
streamlit
prophet
plotly
pandas
paho-mqtt
influxdb-client
python-dotenv